Optimize runtime representation for callable reference subclasses
Instead of generating overrides for getOwner/getName/getSignature in
each anonymous class representing a callable reference, pass them to the
superclass' constructor and store as fields. This occupies some small
memory but helps to reduce the size of the generated class files, and
will be helpful for adding further runtime information to callable
references, such as information about implicit conversions this
reference has been subject to.

Unbound references are going to be optimized independent of the API
version, since the corresponding superclass constructors are available
since Kotlin 1.0, but the optimization for bound references requires API
version >= 1.4.

 #KT-27362 Fixed